visualise: Lazy import numpy, scipy and matplotlib modules
The structure of the dispatch script is such that it imports each of the
subcommand modules, who in-turn import their dependencies. The visualise
subcommand relies numpy, scipy and matplotlib and these dependencies
cause a large lag when being loaded from disk. Below is a hack
comparison of before and after the change, where "echo 3 >
/proc/sys/vm/drop_caches" was executed before each run.

Before:
    ajeffery@lasair ~/Development/fpos
    $ time ./bin/fpos annotate /dev/null -

    real    0m13.413s
    user    0m0.972s
    sys     0m0.200s

After:
    ajeffery@lasair ~/Development/fpos
    $ time ./bin/fpos annotate /dev/null -

    real    0m2.029s
    user    0m0.096s
    sys     0m0.029s