strgrp: Rework main() to cluster a stream of strings
main() now makes less assumptions about the form of the input data.
Previously it assumed the provided data was in a file and that the file
matched the IR of fpos. To ease experimentation with strgrp it now
treats each line as an input string, after trimming the newline(s).