strgrp: strgrp_bin_for() parallelised with OpenMP
The implementation is quite hackish, but shows around a 2x increase with
no attempt to improve the implementation.

The limitation is the use of CCAN's linked list implementation, which
probably won't interact well with OpenMP pragmas. A less hackish
implementation would replace the linked list with an appropriate
realloc() approach so the data is already in a form fit for
parallelisation. However, the improvement even with these hacks is good
enough that the quick-and-dirty implementation is useful.

Some graphs and runtime data are included in the commit to demonstrate
the change in approach, though it's not a completely clean comparison:
The graphs show the implementation with and without the OpenMP pragma.
Both graphs include the overhead of conversion from a linked list to
contiguous memory.