transform: Auto sense input format
Auto sense can be disabled by specifying --form with any of the old
required form arguments. The default value for --form is 'auto'.
Alternatively the auto-sensed form can be confirmed by supplying
--confirm as an option.

A bit can be done to clean up the implementation of transform. The
collection of procedures can be turned into classes to associate the
sense information with each transform option.