visualise: Fix mean daily spend estimate in xy_progressive_mean
The error was that ProgressiveMean.tail() returns the completed months
from the mean member, which contains the per-day progressive mean value
for each month. Summing these values and taking the mean does not give
the intended value. Instead, the last day of the month represents the
mean daily spend for the month. So instead, take the mean of the last
day values to produce a mean value across months.