bin: Add fposx for OSX install/execution
bin/fposx manages the installation of all the build and runtime
dependencies of fpos. The script also manages the bring-up of a shell
environment in which fpos can be executed, taking into account the those
wonderfully unique behaviours of virtualenvs on OSX when you also want
to use matplotlib.

Really, this is a terrible script that shows no regard for good taste.
I'm just banging the rocks together. Treat the script as an MVP. Better
yet, improve it!

Anyway, if you also have a disregard for good taste, invoke the
following to install on OSX:

    $ curl -O https://raw.githubusercontent.com/amboar/fpos/master/bin/fposx
    $ bash fposx install

Similarly, to run fpos:

    $ bash fposx run
    ...
    (ve) $ fpos update ...
    (ve) $ fpos show ...
    (ve) $ logout