bin/fpos: Use env to detect python3
Use /usr/bin/env so the virtualenv python3 executable is found, along
with the packages. During testing on Ubuntu 15.10 it was observed that
the interpreter path hadn't been rewritten for the virtualenv. Maybe
there was a reason for this but it wasn't immediately obvious - further
tests showed the path was correctly re-written on installation.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>