window: Circular queue for spans relative to last transaction
`fpos show mydb` now displays 12 months of transactions relative to the
month of the most recent transaction. Previously a hacky approach had
been used where the date of the least recent transaction was calculated
relative to the current date. This caused issues if the date of the most
recent transaction was more than 12 months in the past relative to the
current date.

The change alters the order of window()'s arguments which leads to a bit
of churn across a few files.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>