Makefile: pip and install targets use --user flags
It turns out that using pip system-wide can hose your system or at least
corrupt your package manager's view of the world on some Linux
distributions.  So, don't install packages system wide.

As a consequence, the install-user make target is gone in favour of
the install target doing as much. Targets still exist for installing
into system directories in a virtualenv, as this is still safe to do.

The README is updated to reflect the new world order.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>